ci: diff-scope the clang-format gate like governance G12 (#285)
* ci: diff-scope the clang-format gate like governance G12

The format job ran clang-format --dry-run over every tracked file on
every push/PR, so it failed regardless of what a PR actually touched
(master itself currently has ~5k pre-existing violations). Scope it to
files changed vs the PR/push base, matching the pattern repo_hygiene.py
already uses for G12.

* governance: regenerate stale ADR index (G9)

INDEX.md hadn't been regenerated since ADR-0031/0032 landed, so G9
failed on every PR/push regardless of content. Re-ran --write-index.

* governance: regenerate drifted z80 capsule (G2)

CAPSULE.md hadn't been regenerated since z80.hpp gained ei_pending/
nmi_accept_count/irq_accept_count, so G2 failed on every PR/push
regardless of content. Re-ran gen_capsule.py.

* m6803: drop tautological opcode<=0xFF check (-Werror=type-limits)

opcode is uint8_t, so opcode <= 0xFFU is always true and GCC's
-Werror=type-limits fails the build. The range collapses to
opcode >= 0xF0U with identical semantics.

* m6803: apply pinned clang-format (pre-existing drift)

File now triggers the diff-scoped format gate (it's in this PR's
diff); the pre-existing line-wrap drift was unrelated to the type-
limits fix but blocks the gate either way. Formatting only, no
behavior change.

* Relax -Wmissing-field-initializers; fix remaining -Werror breaks

The Linux GCC build (-Werror) has apparently never run clean on
master: a local -k0 build surfaced ~100 missing-field-initializer
errors (test/options structs gained fields; call sites deliberately
only set the relevant ones - idiomatic partial aggregate-init, not a
bug) plus a handful of real small issues, fixed here:

- Wno-missing-field-initializers added to the shared GCC/Clang
  warning flags (MSVC unaffected; it doesn't warn on this).
- irem_m78: drop unused expand_3bit (same dead-helper pattern as the
  earlier irem_m52::in_range fix).
- irem_m82 adapter/system tests: narrowing ternary (unsigned int ->
  uint8_t) in identified_m82_source rank init, fixed with an explicit
  cast.
- amiga_adapter: drop unused old_state local (old_ports, which it was
  sliced from, is still used below for keyboard-edge routing).
- Full pinned-clang-format pass on every file touched here, since
  each carried pre-existing line-wrap drift unrelated to these fixes.

---------

+++ src/manifests/irem_m82/tests/m82_system_test.cpp
@@ -227,8 +227,9 @@ namespace {
             if (!known.contains(set_id)) {
                 return std::nullopt;
             }
-            return identified_m82_source{.set_id = set_id,
-                                         .rank = set_id == "dkgensanm82" ? 0U : 2U};
+            return identified_m82_source{
+                .set_id = set_id,
+                .rank = static_cast<std::uint8_t>(set_id == "dkgensanm82" ? 0U : 2U)};
         }
         if (!std::filesystem::is_regular_file(path, ec) ||
             !ends_with_zip(path.filename().string())) {
@@ -368,8 +369,7 @@ namespace {
         rom[0x0000U] = 0x3EU; // LD A,$F0
         rom[0x0001U] = 0xF0U;
         rom[0x0002U] = 0xD3U; // OUT ($82),A
-        rom[0x0003U] =
-            static_cast<std::uint8_t>(mnemos::manifests::irem_m82::z80_port_dac);
+        rom[0x0003U] = static_cast<std::uint8_t>(mnemos::manifests::irem_m82::z80_port_dac);
         rom[0x0004U] = 0x76U; // HALT
         return rom;
     }
@@ -467,9 +467,9 @@ TEST_CASE("m82 checked-in game manifests parse and cover local M82 corpus", "[m8
         declarations.emplace(std::move(set_name), std::move(*parsed.value));
     }
 
-    const std::set<std::string, std::less<>> expected_names{
-        "airduel", "airduelu", "dkgensanm82", "majtitle", "majtitlej",
-        "rtype2",  "rtype2j",  "rtype2jc",  "rtype2m82b"};
+    const std::set<std::string, std::less<>> expected_names{"airduel",  "airduelu",  "dkgensanm82",
+                                                            "majtitle", "majtitlej", "rtype2",
+                                                            "rtype2j",  "rtype2jc",  "rtype2m82b"};
     std::set<std::string, std::less<>> names;
     for (const auto& [set_name, raw_decl] : declarations) {
         INFO("set=" << set_name);
